Rename debug_deserializer to text_deserializer
